+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Added DT asset provider configuration on start up.
 - Use common DTO in all submodels.
 - Use policy hub model in all submodels.
+- Add controller interface for download data.
  
 
 ### Fixed